pr-fix: move web api test fixture types to diff files
